refactor: :recycle: 同义重构：以「推理器」为中心
从「记忆区」迁移重构诸多函数到「推理器」中，基本只把「记忆区」当作「高级概念袋」用；移除多个袋的方法覆盖，预备删除合并

+++ src/nars/inference/BudgetFunctions.java
@@ -248,7 +248,7 @@ public final class BudgetFunctions extends UtilityFunctions {
      * @param forgetRate        The budget for the new item
      * @param relativeThreshold The relative threshold of the bag
      */
-    public static void forget(Budget budgetToBeForget, float forgetRate, float relativeThreshold) {
+    public static void forget(Budget budgetToBeForget, int forgetRate, float relativeThreshold) {
         double quality = budgetToBeForget.getQuality() * relativeThreshold; // re-scaled quality
         final double p = budgetToBeForget.getPriority() - quality; // priority above quality
         if (p > 0) {
